refactor(types): more explicit methods on `ListValue` and `ListRef`
