+++ src/common/src/array/map_array.rs
@@ -337,10 +337,6 @@ mod scalar {
             self.iter().map(|(k, _v)| k).collect()
         }
 
-        pub fn to_owned(self) -> MapValue {
-            MapValue(self.0.to_owned())
-        }
-
         pub fn len(&self) -> usize {
             self.0.len()
         }
