+++ src/common/src/types/scalar_impl.rs
@@ -82,15 +82,6 @@ impl Scalar for StructValue {
     }
 }
 
-/// Implement `Scalar` for `ListValue`.
-impl Scalar for ListValue {
-    type ScalarRefType<'a> = ListRef<'a>;
-
-    fn as_scalar_ref(&self) -> ListRef<'_> {
-        self.into()
-    }
-}
-
 /// Implement `ScalarRef` for `Box<str>`.
 /// `Box<str>` could be converted to `&str`.
 impl<'a> ScalarRef<'a> for &'a str {
@@ -303,18 +294,7 @@ impl<'a> ScalarRef<'a> for StructRef<'a> {
     }
 }
 
-/// Implement `Scalar` for `ListValue`.
-impl<'a> ScalarRef<'a> for ListRef<'a> {
-    type ScalarType = ListValue;
 
-    fn to_owned_scalar(&self) -> ListValue {
-        (*self).into()
-    }
-
-    fn hash_scalar<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.hash_scalar_inner(state)
-    }
-}
 
 impl ScalarImpl {
     pub fn get_ident(&self) -> &'static str {
