+++ src/expr/impl/src/scalar/array.rs
@@ -64,8 +64,9 @@ fn map_from_entries_type_infer(args: &[DataType]) -> Result<DataType, ExprError>
     "map_from_key_values(anyarray, anyarray) -> anymap",
     type_infer = "map_from_key_values_type_infer"
 )]
-fn map_from_key_values(key: ListRef<'_>, value: ListRef<'_>) -> Result<MapValue, ExprError> {
-    MapValue::try_from_kv(key.to_owned(), value.to_owned()).map_err(ExprError::Custom)
+fn map_from_key_values(keys: ListRef<'_>, values: ListRef<'_>) -> Result<MapValue, ExprError> {
+    MapValue::try_from_kv(keys.to_owned_scalar(), values.to_owned_scalar())
+        .map_err(ExprError::Custom)
 }
 
 #[function(
@@ -73,7 +74,7 @@ fn map_from_key_values(key: ListRef<'_>, value: ListRef<'_>) -> Result<MapValue,
     type_infer = "map_from_entries_type_infer"
 )]
 fn map_from_entries(entries: ListRef<'_>) -> Result<MapValue, ExprError> {
-    MapValue::try_from_entries(entries.to_owned()).map_err(ExprError::Custom)
+    MapValue::try_from_entries(entries.to_owned_scalar()).map_err(ExprError::Custom)
 }
 
 /// # Example
@@ -204,7 +205,7 @@ fn map_length<T: TryFrom<usize>>(map: MapRef<'_>) -> Result<T, ExprError> {
 fn map_cat(m1: Option<MapRef<'_>>, m2: Option<MapRef<'_>>) -> Result<Option<MapValue>, ExprError> {
     match (m1, m2) {
         (None, None) => Ok(None),
-        (Some(m), None) | (None, Some(m)) => Ok(Some(m.to_owned())),
+        (Some(m), None) | (None, Some(m)) => Ok(Some(m.to_owned_scalar())),
         (Some(m1), Some(m2)) => Ok(Some(MapValue::concat(m1, m2))),
     }
 }
@@ -233,7 +234,7 @@ fn map_insert(
     value: Option<ScalarRefImpl<'_>>,
 ) -> MapValue {
     let Some(key) = key else {
-        return map.to_owned();
+        return map.to_owned_scalar();
     };
     MapValue::insert(map, key.into_scalar_impl(), value.to_owned_datum())
 }
@@ -258,7 +259,7 @@ fn map_insert(
 #[function("map_delete(anymap, any) -> anymap")]
 fn map_delete(map: MapRef<'_>, key: Option<ScalarRefImpl<'_>>) -> MapValue {
     let Some(key) = key else {
-        return map.to_owned();
+        return map.to_owned_scalar();
     };
     MapValue::delete(map, key)
 }
@@ -314,7 +315,7 @@ fn map_values(map: MapRef<'_>) -> ListValue {
     }"
 )]
 fn map_entries(map: MapRef<'_>) -> ListValue {
-    map.into_inner().to_owned()
+    map.into_inner().to_owned_scalar()
 }
 
 #[cfg(test)]
